+++ parquet/src/file/writer.rs
@@ -993,8 +993,6 @@ mod tests {
 
     #[cfg(feature = "arrow")]
     use arrow_array::RecordBatchReader;
-    use arrow_array::{Int32Array, RecordBatch};
-    use arrow_schema::{Field, Schema};
     use bytes::Bytes;
     use std::fs::File;
 
@@ -1009,7 +1007,6 @@ mod tests {
     use crate::column::reader::get_typed_column_reader;
     use crate::compression::{create_codec, Codec, CodecOptionsBuilder};
     use crate::data_type::{BoolType, ByteArrayType, Int32Type};
-    use crate::file::page_encoding_stats::PageEncodingStats;
     use crate::file::page_index::index::Index;
     use crate::file::properties::EnabledStatistics;
     use crate::file::serialized_reader::ReadOptionsBuilder;
@@ -2399,49 +2396,5 @@ mod tests {
             };
             start += 1;
         }
-    }
-
-    #[test]
-    fn test_page_encoding_statistics_roundtrip() {
-        let batch_schema = Schema::new(vec![Field::new(
-            "int32",
-            arrow_schema::DataType::Int32,
-            false,
-        )]);
-
-        let batch = RecordBatch::try_new(
-            Arc::new(batch_schema.clone()),
-            vec![Arc::new(Int32Array::from(vec![1, 2, 3, 4])) as _],
-        )
-        .unwrap();
-
-        let mut file: File = tempfile::tempfile().unwrap();
-        let mut writer = ArrowWriter::try_new(&mut file, Arc::new(batch_schema), None).unwrap();
-        writer.write(&batch).unwrap();
-        let file_metadata = writer.close().unwrap();
-
-        assert_eq!(file_metadata.row_groups.len(), 1);
-        assert_eq!(file_metadata.row_groups[0].columns.len(), 1);
-        let chunk_meta = file_metadata.row_groups[0].columns[0]
-            .meta_data
-            .as_ref()
-            .expect("column metadata missing");
-        assert!(chunk_meta.encoding_stats.is_some());
-        let chunk_page_stats = chunk_meta.encoding_stats.as_ref().unwrap();
-
-        // check that the read metadata is also correct
-        let options = ReadOptionsBuilder::new().with_page_index().build();
-        let reader = SerializedFileReader::new_with_options(file, options).unwrap();
-
-        let rowgroup = reader.get_row_group(0).expect("row group missing");
-        assert_eq!(rowgroup.num_columns(), 1);
-        let column = rowgroup.metadata().column(0);
-        assert!(column.page_encoding_stats().is_some());
-        let file_page_stats = column.page_encoding_stats().unwrap();
-        let chunk_stats: Vec<PageEncodingStats> = chunk_page_stats
-            .iter()
-            .map(|x| crate::file::page_encoding_stats::try_from_thrift(x).unwrap())
-            .collect();
-        assert_eq!(&chunk_stats, file_page_stats);
     }
 }
